+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 * Unreleased
     * Remove 'class Print' forward declaration that breaks megaAVR like Nano
       Every.
+    * Mark non-virtual const functions with `const`, which preserves backwards
+      compatibility.
 * 1.5 (2020-06-27)
     * Add `LadderButtonConfig` class to support multiple buttons on a single
       analog pin using `analogRead()`. Add documentation in